[draft] cclp

+++ MMOCoreORB/src/server/zone/objects/building/tasks/ScheduleShuttleTask.h
@@ -55,22 +55,28 @@ public:
 			planetManager->scheduleShuttle(strongReference, PlanetManager::SHUTTLEPORT);
 
 		} else {
-			Reference<PlanetTravelPoint*> ptp = planetManager->getNearestPlanetTravelPoint(strongReference, 128.f);
+			if (strongReference->isCityControlShuttle()) {
+				printf( " Schedule Shuttle Task is calling getCityControlLandingPoint \n");
 
-			if (ptp != nullptr && !ptp->isCityControlLandingPoint()) {
-				auto oldShuttle = ptp->getShuttle();
+				Reference<CityControlLandingPoint*> cclp = planetManager->getCityControlLandingPoint(strongReference, 128.f);
+			} else {
+				Reference<PlanetTravelPoint*> ptp = planetManager->getNearestPlanetTravelPoint(strongReference, 128.f);
 
-				if (oldShuttle == nullptr) {
-					if (ptp->isInterplanetary())
-						planetManager->scheduleShuttle(strongReference, PlanetManager::STARPORT);
-					else
-						planetManager->scheduleShuttle(strongReference, PlanetManager::SHUTTLEPORT);
+				if (ptp != nullptr && !ptp->isCityControlLandingPoint()) {
+					auto oldShuttle = ptp->getShuttle();
 
-					ptp->setShuttle(strongReference);
+					if (oldShuttle == nullptr) {
+						if (ptp->isInterplanetary())
+							planetManager->scheduleShuttle(strongReference, PlanetManager::STARPORT);
+						else
+							planetManager->scheduleShuttle(strongReference, PlanetManager::SHUTTLEPORT);
 
-				} else if (oldShuttle != strongReference) {
-					strongReference->destroyObjectFromWorld(true);
-					strongReference->destroyObjectFromDatabase(true);
+						ptp->setShuttle(strongReference);
+
+					} else if (oldShuttle != strongReference) {
+						strongReference->destroyObjectFromWorld(true);
+						strongReference->destroyObjectFromDatabase(true);
+					}
 				}
 			}
 		}
